requirements.txt
reme-ai[light]>=0.3.0
openai>=1.30.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
prompt_toolkit>=3.0.0

# union-search-skill 基础依赖（no_api_key 分组所需）
requests>=2.31.0
lxml>=5.0.0

# browser-use: AI 驱动的浏览器自动化
browser-use>=0.12.0
langchain-openai>=0.3.0

# playwright 是 browser-use 的底层依赖
# 安装后需运行: playwright install chromium
playwright>=1.40.0

# gateway 模式：FastAPI + uvicorn 异步 HTTP 服务
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
